requirements.txt
# CourseForge 依赖包列表

# Web 框架
streamlit>=1.28.0

# AI API 客户端
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.4.0
requests>=2.28.0

# Office 文档处理
python-pptx>=0.6.21
PyPDF2>=3.0.0
python-docx>=1.1.0

# 数据处理
pandas>=2.0.0

# 可选：进度条美化
tqdm>=4.65.0
